# MM-7633: Optimize memory utilization during file uploads