--- a/api4/apitestlib.go
+++ b/api4/apitestlib.go
@@ -573,103 +573,68 @@ func CheckNoError(t *testing.T, resp *model.Response) {
 	t.Helper()
 
 	if resp.Error != nil {
-		t.Fatal("Expected no error, got " + resp.Error.Error())
+		t.Fatalf("Expected no error, got %q", resp.Error.Error())
 	}
 }
 
-func CheckCreatedStatus(t *testing.T, resp *model.Response) {
+func checkHTTPStatus(t *testing.T, resp *model.Response, status int, expectError bool) {
 	t.Helper()
-
-	if resp.StatusCode != http.StatusCreated {
-		t.Log("actual: " + strconv.Itoa(resp.StatusCode))
-		t.Log("expected: " + strconv.Itoa(http.StatusCreated))
-		t.Fatal("wrong status code")
-	}
-}
-
-func CheckForbiddenStatus(t *testing.T, resp *model.Response) {
-	t.Helper()
-
-	if resp.Error == nil {
-		t.Fatal("should have errored with status:" + strconv.Itoa(http.StatusForbidden))
-		return
+	if expectError {
+		if resp.Error == nil {
+			t.Fatalf("Expected an error with status:%v, got nil", status)
+		}
+	} else {
+		CheckNoError(t, resp)
 	}
 
-	if resp.StatusCode != http.StatusForbidden {
-		t.Log("actual: " + strconv.Itoa(resp.StatusCode))
-		t.Log("expected: " + strconv.Itoa(http.StatusForbidden))
-		t.Fatal("wrong status code")
-	}
-}
-
-func CheckUnauthorizedStatus(t *testing.T, resp *model.Response) {
-	t.Helper()
-
-	if resp.Error == nil {
-		t.Fatal("should have errored with status:" + strconv.Itoa(http.StatusUnauthorized))
-		return
-	}
-
-	if resp.StatusCode != http.StatusUnauthorized {
-		t.Log("actual: " + strconv.Itoa(resp.StatusCode))
-		t.Log("expected: " + strconv.Itoa(http.StatusUnauthorized))
-		t.Fatal("wrong status code")
-	}
-}
-
-func CheckNotFoundStatus(t *testing.T, resp *model.Response) {
-	t.Helper()
-
-	if resp.Error == nil {
-		t.Fatal("should have errored with status:" + strconv.Itoa(http.StatusNotFound))
-		return
-	}
-
-	if resp.StatusCode != http.StatusNotFound {
-		t.Log("actual: " + strconv.Itoa(resp.StatusCode))
-		t.Log("expected: " + strconv.Itoa(http.StatusNotFound))
-		t.Fatal("wrong status code")
-	}
-}
-
-func CheckBadRequestStatus(t *testing.T, resp *model.Response) {
-	t.Helper()
-
-	if resp.Error == nil {
-		t.Fatal("should have errored with status:" + strconv.Itoa(http.StatusBadRequest))
-		return
-	}
-
-	if resp.StatusCode != http.StatusBadRequest {
-		t.Log("actual: " + strconv.Itoa(resp.StatusCode))
-		t.Log("expected: " + strconv.Itoa(http.StatusBadRequest))
-		t.Fatal("wrong status code")
-	}
-}
-
-func CheckNotImplementedStatus(t *testing.T, resp *model.Response) {
-	t.Helper()
-
-	if resp.Error == nil {
-		t.Fatal("should have errored with status:" + strconv.Itoa(http.StatusNotImplemented))
-		return
-	}
-
-	if resp.StatusCode != http.StatusNotImplemented {
-		t.Log("actual: " + strconv.Itoa(resp.StatusCode))
-		t.Log("expected: " + strconv.Itoa(http.StatusNotImplemented))
-		t.Fatal("wrong status code")
+	if resp.StatusCode != status {
+		t.Fatalf("Expected status code %v, got %v", status, resp.StatusCode)
 	}
 }
 
 func CheckOKStatus(t *testing.T, resp *model.Response) {
 	t.Helper()
+	checkHTTPStatus(t, resp, http.StatusOK, false)
+}
 
-	CheckNoError(t, resp)
+func CheckCreatedStatus(t *testing.T, resp *model.Response) {
+	t.Helper()
+	checkHTTPStatus(t, resp, http.StatusCreated, false)
+}
 
-	if resp.StatusCode != http.StatusOK {
-		t.Fatalf("wrong status code. expected %d got %d", http.StatusOK, resp.StatusCode)
-	}
+func CheckForbiddenStatus(t *testing.T, resp *model.Response) {
+	t.Helper()
+	checkHTTPStatus(t, resp, http.StatusForbidden, true)
+}
+
+func CheckUnauthorizedStatus(t *testing.T, resp *model.Response) {
+	t.Helper()
+	checkHTTPStatus(t, resp, http.StatusUnauthorized, true)
+}
+
+func CheckNotFoundStatus(t *testing.T, resp *model.Response) {
+	t.Helper()
+	checkHTTPStatus(t, resp, http.StatusNotFound, true)
+}
+
+func CheckBadRequestStatus(t *testing.T, resp *model.Response) {
+	t.Helper()
+	checkHTTPStatus(t, resp, http.StatusBadRequest, true)
+}
+
+func CheckNotImplementedStatus(t *testing.T, resp *model.Response) {
+	t.Helper()
+	checkHTTPStatus(t, resp, http.StatusNotImplemented, true)
+}
+
+func CheckRequestEntityTooLargeStatus(t *testing.T, resp *model.Response) {
+	t.Helper()
+	checkHTTPStatus(t, resp, http.StatusRequestEntityTooLarge, true)
+}
+
+func CheckInternalErrorStatus(t *testing.T, resp *model.Response) {
+	t.Helper()
+	checkHTTPStatus(t, resp, http.StatusInternalServerError, true)
 }
 
 func CheckErrorMessage(t *testing.T, resp *model.Response, errorId string) {
@@ -684,21 +649,6 @@ func CheckErrorMessage(t *testing.T, resp *model.Response, errorId string) {
 		t.Log("actual: " + resp.Error.Id)
 		t.Log("expected: " + errorId)
 		t.Fatal("incorrect error message")
-	}
-}
-
-func CheckInternalErrorStatus(t *testing.T, resp *model.Response) {
-	t.Helper()
-
-	if resp.Error == nil {
-		t.Fatal("should have errored with status:" + strconv.Itoa(http.StatusInternalServerError))
-		return
-	}
-
-	if resp.StatusCode != http.StatusInternalServerError {
-		t.Log("actual: " + strconv.Itoa(resp.StatusCode))
-		t.Log("expected: " + strconv.Itoa(http.StatusInternalServerError))
-		t.Fatal("wrong status code")
 	}
 }