@levb 2018-11-21
@hmhealey It is my understanding that we upload the original image "as is", we do not flip it since it already contains the necessary metadata, and all user agents respect it. However, if the image was flipped, we seem to be recording its dimensions in FileInfo to match the actual display.

 | size | orientation | FileInfo size
-------|------|------|-----
uploaded image | WxH | 5 (90º?) | n/a
stored image | WxH | 5 (90º?) | HxW
thumb | calculated, from HxW | 1 (normal) | n/a
preview | calculated, from HxW | 1 (normal) | n/a

I figured that perhaps we were using the FileInfo dimensions for sizing UI elements before the image is downloaded, so they had to be "real"? If you can point me at where/how they are used, maybe we can add a test from that?