@levb 2018-11-26
@hmhealey I think we should leave it as is for now, and I'll add a test that documents/verifies the behavior. When we look into cleaning up image processing performance issues, we can look into maybe adding the thumbnail/preview sizes to FileInfo explicitly, that way the client doesn't need to guess-calculate them.