@levb 2018-12-03
@lieut-data I appreciate the advantages of simplifying the new API by keeping the backwards compatibility out of it. The main reason I decided to include the "legacy" path in the new implementation so that we can cleanly switch the various file upload use-cases to the new API and replace all of the old code. It was also nice for testing not to have 2 separate end-points to test.

With the changes you hinted at, I replaced the obscure request hacking with a couple small functions to make our own `MultipartReader` and I think it should be easier to read now. Let me know if you still feel reserved about the backwards compatibility issue.